requirements.txt
torch==2.5.1
transformers==4.47.1
pandas==2.2.3
numpy==1.26.4
tqdm==4.67.0
scipy==1.14.1
scikit-learn==1.5.2
nltk==3.9.1
peft==0.13.2
monai==1.4.0
lighter_zoo==0.1.3
torchvision==0.20.1
smart_open==7.0.5
boto3==1.35.58
torch_geometric==2.6.1
python-dotenv==1.0.1
einops==0.8.0
sentencepiece==0.2.0
tensorboard==2.18.0
rouge-score==0.1.2